requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==1.11
django-filter==1.1.0
django-filters==0.2.1
django-widget-tweaks==1.4.2
djangorestframework==3.6.2
drfdocs==0.0.11
gunicorn==19.6.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
pytz==2018.4
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
